requirements-test.txt
# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
requests-mock>=1.11.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0